Fix JSON parsing bug introduced by accidental typo
Using a comma instead of a period caused a string to be passed to md5() which was then interpreted as true. Passing true as the second parameter to md5() causes it to return the raw binary digest, rather than a string representation of the digest which in turn caused a JSON parsing error.